SpreadsheetExpressionReferenceContext extracted from SpreadsheetExpressionEvaluationContext
